requirements.txt
psycopg2-binary
sqlalchemy
PyYAML
geoalchemy2
geopandas
loguru
shapely
pandas